+++ src/LinkContext.cs
@@ -1,0 +1,8 @@
+namespace Tailwind.Css.TagHelpers;
+
+public class LinkContext
+{
+    public bool IsMatch { get; set; }
+
+    public PathMatchStyle MatchStyle { get; set; }
+}